+++ com.unity.netcode.gameobjects/Runtime/Components/NetworkAnimator.cs
@@ -1,6 +1,7 @@
 #if COM_UNITY_MODULES_ANIMATION
 using System;
 using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
 using Unity.Netcode.Runtime;
@@ -33,7 +34,7 @@ namespace Unity.Netcode.Components
                 }
                 else
                 {
-                    m_NetworkAnimator.SendAnimStateClientRpc(animationUpdate.AnimationMessage, animationUpdate.ClientRpcParams);
+                    m_NetworkAnimator.SendClientAnimStateRpc(animationUpdate.AnimationMessage, animationUpdate.RpcParams);
                 }
             }
 
@@ -47,7 +48,7 @@ namespace Unity.Netcode.Components
                 }
                 else
                 {
-                    m_NetworkAnimator.SendParametersUpdateClientRpc(sendEntry.ParametersUpdateMessage, sendEntry.ClientRpcParams);
+                    m_NetworkAnimator.SendClientParametersUpdateRpc(sendEntry.ParametersUpdateMessage, sendEntry.RpcParams);
                 }
             }
             m_SendParameterUpdates.Clear();
@@ -62,15 +63,22 @@ namespace Unity.Netcode.Components
                 {
                     if (!sendEntry.SendToServer)
                     {
-                        m_NetworkAnimator.SendAnimTriggerClientRpc(sendEntry.AnimationTriggerMessage, sendEntry.ClientRpcParams);
+                        m_NetworkAnimator.SendClientAnimTriggerRpc(sendEntry.AnimationTriggerMessage, sendEntry.RpcParams);
                     }
                     else
                     {
-                        m_NetworkAnimator.SendAnimTriggerServerRpc(sendEntry.AnimationTriggerMessage);
+                        m_NetworkAnimator.SendServerAnimTriggerRpc(sendEntry.AnimationTriggerMessage);
                     }
                 }
             }
             m_SendTriggerUpdates.Clear();
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private bool HasAuthority()
+        {
+            var isServerAuthority = m_NetworkAnimator.IsServerAuthoritative();
+            return (!isServerAuthority && m_NetworkAnimator.IsOwner) || (isServerAuthority && (m_NetworkAnimator.IsServer || m_NetworkAnimator.IsOwner));
         }
 
         /// <inheritdoc />
@@ -80,25 +88,32 @@ namespace Unity.Netcode.Components
             {
                 case NetworkUpdateStage.PreUpdate:
                     {
-                        // Only the owner or the server send messages
-                        if (m_NetworkAnimator.IsOwner || m_IsServer)
+                        // NOTE: This script has an order of operations requirement where
+                        // the authority and/or server will flush messages first, parameter updates are applied
+                        // for all instances, and then only the authority will check for animator changes. Changing
+                        // the order could cause timing related issues.
+
+                        var hasAuthority = HasAuthority();
+                        // Only the authority or the server will send messages
+                        if (hasAuthority || m_IsServer)
                         {
                             // Flush any pending messages
                             FlushMessages();
                         }
 
                         // Everyone applies any parameters updated
-                        for (int i = 0; i < m_ProcessParameterUpdates.Count; i++)
+                        if (m_ProcessParameterUpdates.Count > 0)
                         {
-                            var parameterUpdate = m_ProcessParameterUpdates[i];
-                            m_NetworkAnimator.UpdateParameters(ref parameterUpdate);
+                            for (int i = 0; i < m_ProcessParameterUpdates.Count; i++)
+                            {
+                                var parameterUpdate = m_ProcessParameterUpdates[i];
+                                m_NetworkAnimator.UpdateParameters(ref parameterUpdate);
+                            }
+                            m_ProcessParameterUpdates.Clear();
                         }
-                        m_ProcessParameterUpdates.Clear();
-                        var isServerAuthority = m_NetworkAnimator.IsServerAuthoritative();
 
-                        // owners when owner authoritative or the server when server authoritative are the only instances that
-                        // checks for Animator changes
-                        if ((!isServerAuthority && m_NetworkAnimator.IsOwner) || (isServerAuthority && m_NetworkAnimator.IsServer))
+                        // Only the authority checks for Animator changes
+                        if (hasAuthority)
                         {
                             m_NetworkAnimator.CheckForAnimatorChanges();
                         }
@@ -112,7 +127,7 @@ namespace Unity.Netcode.Components
         /// </summary>
         private struct AnimationUpdate
         {
-            public ClientRpcParams ClientRpcParams;
+            public RpcParams RpcParams;
             public NetworkAnimator.AnimationMessage AnimationMessage;
         }
 
@@ -121,14 +136,14 @@ namespace Unity.Netcode.Components
         /// <summary>
         /// Invoked when a server needs to forwarding an update to the animation state
         /// </summary>
-        internal void SendAnimationUpdate(NetworkAnimator.AnimationMessage animationMessage, ClientRpcParams clientRpcParams = default)
+        internal void SendAnimationUpdate(NetworkAnimator.AnimationMessage animationMessage, RpcParams rpcParams = default)
         {
-            m_SendAnimationUpdates.Add(new AnimationUpdate() { ClientRpcParams = clientRpcParams, AnimationMessage = animationMessage });
+            m_SendAnimationUpdates.Add(new AnimationUpdate() { RpcParams = rpcParams, AnimationMessage = animationMessage });
         }
 
         private struct ParameterUpdate
         {
-            public ClientRpcParams ClientRpcParams;
+            public RpcParams RpcParams;
             public NetworkAnimator.ParametersUpdateMessage ParametersUpdateMessage;
         }
 
@@ -137,9 +152,9 @@ namespace Unity.Netcode.Components
         /// <summary>
         /// Invoked when a server needs to forwarding an update to the parameter state
         /// </summary>
-        internal void SendParameterUpdate(NetworkAnimator.ParametersUpdateMessage parametersUpdateMessage, ClientRpcParams clientRpcParams = default)
+        internal void SendParameterUpdate(NetworkAnimator.ParametersUpdateMessage parametersUpdateMessage, RpcParams rpcParams = default)
         {
-            m_SendParameterUpdates.Add(new ParameterUpdate() { ClientRpcParams = clientRpcParams, ParametersUpdateMessage = parametersUpdateMessage });
+            m_SendParameterUpdates.Add(new ParameterUpdate() { RpcParams = rpcParams, ParametersUpdateMessage = parametersUpdateMessage });
         }
 
         private List<NetworkAnimator.ParametersUpdateMessage> m_ProcessParameterUpdates = new List<NetworkAnimator.ParametersUpdateMessage>();
@@ -151,7 +166,7 @@ namespace Unity.Netcode.Components
         private struct TriggerUpdate
         {
             public bool SendToServer;
-            public ClientRpcParams ClientRpcParams;
+            public RpcParams RpcParams;
             public NetworkAnimator.AnimationTriggerMessage AnimationTriggerMessage;
         }
 
@@ -160,9 +175,9 @@ namespace Unity.Netcode.Components
         /// <summary>
         /// Invoked when a server needs to forward an update to a Trigger state
         /// </summary>
-        internal void QueueTriggerUpdateToClient(NetworkAnimator.AnimationTriggerMessage animationTriggerMessage, ClientRpcParams clientRpcParams = default)
+        internal void QueueTriggerUpdateToClient(NetworkAnimator.AnimationTriggerMessage animationTriggerMessage, RpcParams clientRpcParams = default)
         {
-            m_SendTriggerUpdates.Add(new TriggerUpdate() { ClientRpcParams = clientRpcParams, AnimationTriggerMessage = animationTriggerMessage });
+            m_SendTriggerUpdates.Add(new TriggerUpdate() { RpcParams = clientRpcParams, AnimationTriggerMessage = animationTriggerMessage });
         }
 
         internal void QueueTriggerUpdateToServer(NetworkAnimator.AnimationTriggerMessage animationTriggerMessage)
@@ -190,6 +205,12 @@ namespace Unity.Netcode.Components
     [HelpURL(HelpUrls.NetworkAnimator)]
     public class NetworkAnimator : NetworkBehaviour, ISerializationCallbackReceiver
     {
+#if UNITY_EDITOR
+        [HideInInspector]
+        [SerializeField]
+        internal bool NetworkAnimatorExpanded;
+#endif
+
         [Serializable]
         internal class TransitionStateinfo
         {
@@ -200,6 +221,49 @@ namespace Unity.Netcode.Components
             public float TransitionDuration;
             public int TriggerNameHash;
             public int TransitionIndex;
+        }
+
+        /// <summary>
+        /// Determines if the server or client owner pushes animation state updates.
+        /// </summary>
+        public enum AuthorityModes
+        {
+            /// <summary>
+            /// Server pushes animator state updates.
+            /// </summary>
+            Server,
+            /// <summary>
+            /// Client owner pushes animator state updates.
+            /// </summary>
+            Owner,
+        }
+
+        /// <summary>
+        /// Determines whether this <see cref="NetworkAnimator"/> instance will have state updates pushed by the server or the client owner.
+        /// <see cref="AuthorityModes"/>
+        /// </summary>
+#if MULTIPLAYER_SERVICES_SDK_INSTALLED
+        [Tooltip("Selects who has authority(sends state updates) over the<see cref=\"NetworkAnimator\"/> instance.When the network topology is set to distributed authority, this always defaults to owner authority.If server (the default), then only server-side adjustments to the " +
+            "<see cref=\"NetworkAnimator\"> instance will be synchronized with clients. If owner (or client), then only the owner-side adjustments to the <see cref=\"NetworkAnimator\"/> instance will be synchronized with both the server and other clients.")]
+#else
+        [Tooltip("Selects who has authority (sends state updates) over the <see cref=\"NetworkAnimator\"/> instance. If server (the default), then only server-side adjustments to the <see cref=\"NetworkAnimator\"/> instance will be synchronized with clients. If owner (or client), " +
+            "then only the owner-side adjustments to the <see cref=\"NetworkAnimator\"/> instance will be synchronized with both the server and other clients.")]
+#endif
+        public AuthorityModes AuthorityMode;
+
+        [Tooltip("The animator that this NetworkAnimator component will be synchronizing.")]
+        [SerializeField] private Animator m_Animator;
+
+        /// <summary>
+        /// The <see cref="Animator"/> associated with this <see cref="NetworkAnimator"/> instance.
+        /// </summary>
+        public Animator Animator
+        {
+            get { return m_Animator; }
+            set
+            {
+                m_Animator = value;
+            }
         }
 
         /// <summary>
@@ -237,8 +301,36 @@ namespace Unity.Netcode.Components
             }
         }
 
+        [Serializable]
+        internal class AnimatorParameterEntry
+        {
+#pragma warning disable IDE1006
+            [HideInInspector]
+            public string name;
+#pragma warning restore IDE1006
+            public int NameHash;
+            public bool Synchronize;
+            public AnimatorControllerParameterType ParameterType;
+        }
+
+        [Serializable]
+        internal class AnimatorParametersListContainer
+        {
+            public List<AnimatorParameterEntry> ParameterEntries = new List<AnimatorParameterEntry>();
+        }
+
+        [SerializeField]
+        internal AnimatorParametersListContainer AnimatorParameterEntries;
+
+        internal Dictionary<int, AnimatorParameterEntry> AnimatorParameterEntryTable = new Dictionary<int, AnimatorParameterEntry>();
 
 #if UNITY_EDITOR
+        [HideInInspector]
+        [SerializeField]
+        internal bool AnimatorParametersExpanded;
+
+        internal Dictionary<int, AnimatorControllerParameter> ParameterToNameLookup = new Dictionary<int, AnimatorControllerParameter>();
+
         private void ParseStateMachineStates(int layerIndex, ref AnimatorController animatorController, ref AnimatorStateMachine stateMachine)
         {
             for (int y = 0; y < stateMachine.states.Length; y++)
@@ -329,6 +421,73 @@ namespace Unity.Netcode.Components
             }
         }
 
+        internal void ProcessParameterEntries()
+        {
+            if (!Animator)
+            {
+                if (AnimatorParameterEntries != null && AnimatorParameterEntries.ParameterEntries.Count > 0)
+                {
+                    AnimatorParameterEntries.ParameterEntries.Clear();
+                }
+                return;
+            }
+
+            var parameters = Animator.parameters;
+
+            var parametersToRemove = new List<AnimatorParameterEntry>();
+            ParameterToNameLookup.Clear();
+            foreach (var parameter in parameters)
+            {
+                ParameterToNameLookup.Add(parameter.nameHash, parameter);
+            }
+
+            // Rebuild the parameter entry table for the inspector view
+            AnimatorParameterEntryTable.Clear();
+            foreach (var parameterEntry in AnimatorParameterEntries.ParameterEntries)
+            {
+                // Check for removed parameters.
+                if (!ParameterToNameLookup.ContainsKey(parameterEntry.NameHash))
+                {
+                    parametersToRemove.Add(parameterEntry);
+                    // Skip this removed entry
+                    continue;
+                }
+
+                // Build the list of known parameters
+                if (!AnimatorParameterEntryTable.ContainsKey(parameterEntry.NameHash))
+                {
+                    AnimatorParameterEntryTable.Add(parameterEntry.NameHash, parameterEntry);
+                }
+
+                var parameter = ParameterToNameLookup[parameterEntry.NameHash];
+                parameterEntry.name = parameter.name;
+                parameterEntry.ParameterType = parameter.type;
+            }
+
+            // Update for removed parameters
+            foreach (var parameterEntry in parametersToRemove)
+            {
+                AnimatorParameterEntries.ParameterEntries.Remove(parameterEntry);
+            }
+
+            // Update any newly added parameters
+            foreach (var parameterLookUp in ParameterToNameLookup)
+            {
+                if (!AnimatorParameterEntryTable.ContainsKey(parameterLookUp.Value.nameHash))
+                {
+                    var animatorParameterEntry = new AnimatorParameterEntry()
+                    {
+                        name = parameterLookUp.Value.name,
+                        NameHash = parameterLookUp.Value.nameHash,
+                        ParameterType = parameterLookUp.Value.type,
+                        Synchronize = true,
+                    };
+                    AnimatorParameterEntries.ParameterEntries.Add(animatorParameterEntry);
+                    AnimatorParameterEntryTable.Add(parameterLookUp.Value.nameHash, animatorParameterEntry);
+                }
+            }
+        }
+
         /// <summary>
         /// In-Editor Only
         /// Virtual OnValidate method for custom derived NetworkAnimator classes.
@@ -336,6 +495,7 @@ namespace Unity.Netcode.Components
         protected virtual void OnValidate()
         {
             BuildTransitionStateInfoList();
+            ProcessParameterEntries();
         }
 #endif
 
@@ -489,24 +649,19 @@ namespace Unity.Netcode.Components
             }
         }
 
-        [SerializeField] private Animator m_Animator;
-
-        public Animator Animator
-        {
-            get { return m_Animator; }
-            set
-            {
-                m_Animator = value;
-            }
-        }
-
-        internal bool IsServerAuthoritative()
+        /// <summary>
+        /// Determines whether the <see cref="NetworkAnimator"/> is <see cref="AuthorityModes.Server"/> or <see cref="AuthorityModes.Owner"/> based on the <see cref="AuthorityMode"/> field.
+        /// Optionally, you can still derive from <see cref="NetworkAnimator"/> and override the <see cref="OnIsServerAuthoritative"/> method.
+        /// </summary>
+        /// <returns><see cref="true"/> or <see cref="false"/></returns>
+        public bool IsServerAuthoritative()
         {
             return OnIsServerAuthoritative();
         }
 
         /// <summary>
-        /// Override this method and return false to switch to owner authoritative mode.
+        /// Override this method and return false to switch to owner authoritative mode.<br />
+        /// Alternately, you can update the <see cref="AuthorityMode"/> field within the inspector view to select the authority mode.
         /// </summary>
         /// <remarks>
         /// When using a distributed authority network topology, this will default to
@@ -514,12 +669,11 @@ namespace Unity.Netcode.Components
         /// </remarks>
         protected virtual bool OnIsServerAuthoritative()
         {
-            if (!m_LocalNetworkManager)
+            if (DistributedAuthorityMode)
             {
-                return true;
+                return false;
             }
-
-            return !DistributedAuthorityMode;
+            return AuthorityMode == AuthorityModes.Server;
         }
 
         private int[] m_TransitionHash;
@@ -527,8 +681,8 @@ namespace Unity.Netcode.Components
         private float[] m_LayerWeights;
         private static byte[] s_EmptyArray = new byte[] { };
         private List<int> m_ParametersToUpdate;
-        private List<ulong> m_ClientSendList;
-        private ClientRpcParams m_ClientRpcParams;
+        private RpcParams m_RpcParams;
+        private RpcTargetGroup m_TargetGroup;
         private AnimationMessage m_AnimationMessage;
         private NetworkAnimatorStateChangeHandler m_NetworkAnimatorStateChangeHandler;
 
@@ -539,6 +693,7 @@ namespace Unity.Netcode.Components
 
         private unsafe struct AnimatorParamCache
         {
+            internal bool Exclude;
             internal int Hash;
             internal int Type;
             internal fixed byte Value[4]; // this is a max size of 4 bytes
@@ -583,6 +738,8 @@ namespace Unity.Netcode.Components
         {
             SpawnCleanup();
 
+            m_TargetGroup?.Dispose();
+
             if (m_CachedAnimatorParameters != null && m_CachedAnimatorParameters.IsCreated)
             {
                 m_CachedAnimatorParameters.Dispose();
@@ -603,6 +760,11 @@ namespace Unity.Netcode.Components
                 Debug.LogError($"{nameof(NetworkAnimator)} {name} does not have an {nameof(UnityEngine.Animator)} assigned to it. The {nameof(NetworkAnimator)} will not initialize properly.");
 #endif
                 return;
+            }
+
+            foreach (var parameterEntry in AnimatorParameterEntries.ParameterEntries)
+            {
+                AnimatorParameterEntryTable.TryAdd(parameterEntry.NameHash, parameterEntry);
             }
 
             int layers = m_Animator.layerCount;
@@ -648,11 +810,17 @@ namespace Unity.Netcode.Components
             for (var i = 0; i < parameters.Length; i++)
             {
                 var parameter = parameters[i];
+                var synchronizeParameter = true;
+                if (AnimatorParameterEntryTable.ContainsKey(parameter.nameHash))
+                {
+                    synchronizeParameter = AnimatorParameterEntryTable[parameter.nameHash].Synchronize;
+                }
 
                 var cacheParam = new AnimatorParamCache
                 {
                     Type = UnsafeUtility.EnumToInt(parameter.type),
-                    Hash = parameter.nameHash
+                    Hash = parameter.nameHash,
+                    Exclude = !synchronizeParameter
                 };
 
                 unsafe
@@ -736,12 +904,12 @@ namespace Unity.Netcode.Components
                 NetworkLog.LogWarningServer($"[{gameObject.name}][{nameof(NetworkAnimator)}] {nameof(Animator)} is not assigned! Animation synchronization will not work for this instance!");
             }
 
-            m_ClientSendList = new List<ulong>(128);
-            m_ClientRpcParams = new ClientRpcParams
+            m_TargetGroup = RpcTarget.Group(new List<ulong>(128), RpcTargetUse.Persistent) as RpcTargetGroup;
+            m_RpcParams = new RpcParams()
             {
-                Send = new ClientRpcSendParams
+                Send = new RpcSendParams()
                 {
-                    TargetClientIds = m_ClientSendList
+                    Target = m_TargetGroup
                 }
             };
 
@@ -766,6 +934,10 @@ namespace Unity.Netcode.Components
                 m_ParametersToUpdate.Clear();
                 for (int i = 0; i < m_CachedAnimatorParameters.Length; i++)
                 {
+                    if (m_CachedAnimatorParameters[i].Exclude)
+                    {
+                        continue;
+                    }
                     m_ParametersToUpdate.Add(i);
                 }
                 // Write, apply, and serialize
@@ -1009,27 +1181,27 @@ namespace Unity.Netcode.Components
                 else
                 if (!IsServer && IsOwner)
                 {
-                    SendAnimStateServerRpc(m_AnimationMessage);
+                    SendServerAnimStateRpc(m_AnimationMessage);
                 }
                 else
                 {
                     // Just notify all remote clients and not the local server
-                    m_ClientSendList.Clear();
+                    m_TargetGroup.Clear();
                     foreach (var clientId in m_LocalNetworkManager.ConnectionManager.ConnectedClientIds)
                     {
                         if (clientId == m_LocalNetworkManager.LocalClientId || !NetworkObject.Observers.Contains(clientId))
                         {
                             continue;
                         }
-                        m_ClientSendList.Add(clientId);
+                        m_TargetGroup.Add(clientId);
                     }
-                    m_ClientRpcParams.Send.TargetClientIds = m_ClientSendList;
-                    SendAnimStateClientRpc(m_AnimationMessage, m_ClientRpcParams);
+                    m_RpcParams.Send.Target = m_TargetGroup;
+                    SendClientAnimStateRpc(m_AnimationMessage, m_RpcParams);
                 }
             }
         }
 
-        private void SendParametersUpdate(ClientRpcParams clientRpcParams = default, bool sendDirect = false)
+        private void SendParametersUpdate(RpcParams rpcParams = default, bool sendDirect = false)
         {
             WriteParameters(ref m_ParameterWriter);
 
@@ -1052,17 +1224,17 @@ namespace Unity.Netcode.Components
             {
                 if (!IsServer)
                 {
-                    SendParametersUpdateServerRpc(parametersMessage);
+                    SendServerParametersUpdateRpc(parametersMessage);
                 }
                 else
                 {
                     if (sendDirect)
                     {
-                        SendParametersUpdateClientRpc(parametersMessage, clientRpcParams);
+                        SendClientParametersUpdateRpc(parametersMessage, rpcParams);
                     }
                     else
                     {
-                        m_NetworkAnimatorStateChangeHandler.SendParameterUpdate(parametersMessage, clientRpcParams);
+                        m_NetworkAnimatorStateChangeHandler.SendParameterUpdate(parametersMessage, rpcParams);
                     }
                 }
             }
@@ -1092,6 +1264,11 @@ namespace Unity.Netcode.Components
             for (int i = 0; i < m_CachedAnimatorParameters.Length; i++)
             {
                 ref var cacheValue = ref UnsafeUtility.ArrayElementAsRef<AnimatorParamCache>(m_CachedAnimatorParameters.GetUnsafePtr(), i);
+
+                if (cacheValue.Exclude)
+                {
+                    continue;
+                }
 
                 // If a parameter gets controlled by a curve during runtime after initialization of NetworkAnimator
                 // then ignore changes to this parameter. We are not removing the parameter in the event that
@@ -1149,6 +1326,13 @@ namespace Unity.Netcode.Components
             foreach (var parameterIndex in m_ParametersToUpdate)
             {
                 ref var cacheValue = ref UnsafeUtility.ArrayElementAsRef<AnimatorParamCache>(m_CachedAnimatorParameters.GetUnsafePtr(), parameterIndex);
+
+                if (cacheValue.Exclude)
+                {
+                    Debug.LogWarning($"Parameter hash:{cacheValue.Hash} should be excluded but is in the parameters to update list when writing parameter values!");
+                    continue;
+                }
+
                 var hash = cacheValue.Hash;
                 BytePacker.WriteValuePacked(writer, (uint)parameterIndex);
                 if (cacheValue.Type == AnimationParamEnumWrapper.AnimatorControllerParameterInt)
@@ -1317,8 +1501,8 @@ namespace Unity.Netcode.Components
         /// Server-side animator parameter update request
         /// The server sets its local parameters and then forwards the message to the remaining clients
         /// </summary>
-        [ServerRpc]
-        private unsafe void SendParametersUpdateServerRpc(ParametersUpdateMessage parametersUpdate, ServerRpcParams serverRpcParams = default)
+        [Rpc(SendTo.Server, AllowTargetOverride = true, InvokePermission = RpcInvokePermission.Owner)]
+        private unsafe void SendServerParametersUpdateRpc(ParametersUpdateMessage parametersUpdate, RpcParams rpcParams = default)
         {
             if (IsServerAuthoritative())
             {
@@ -1326,7 +1510,7 @@ namespace Unity.Netcode.Components
             }
             else
             {
-                if (serverRpcParams.Receive.SenderClientId != OwnerClientId)
+                if (rpcParams.Receive.SenderClientId != OwnerClientId)
                 {
                     return;
                 }
@@ -1337,26 +1521,26 @@ namespace Unity.Netcode.Components
                     return;
                 }
 
-                m_ClientSendList.Clear();
+                m_TargetGroup.Clear();
                 foreach (var clientId in connectedClientIds)
                 {
-                    if (clientId == serverRpcParams.Receive.SenderClientId || clientId == NetworkManager.ServerClientId || !NetworkObject.Observers.Contains(clientId))
+                    if (clientId == rpcParams.Receive.SenderClientId || clientId == NetworkManager.ServerClientId || !NetworkObject.Observers.Contains(clientId))
                     {
                         continue;
                     }
-                    m_ClientSendList.Add(clientId);
+                    m_TargetGroup.Add(clientId);
                 }
 
-                m_ClientRpcParams.Send.TargetClientIds = m_ClientSendList;
-                m_NetworkAnimatorStateChangeHandler.SendParameterUpdate(parametersUpdate, m_ClientRpcParams);
+                m_RpcParams.Send.Target = m_TargetGroup;
+                m_NetworkAnimatorStateChangeHandler.SendParameterUpdate(parametersUpdate, m_RpcParams);
             }
         }
 
         /// <summary>
         /// Distributed Authority: Updates the client's animator's parameters
         /// </summary>
-        [Rpc(SendTo.NotAuthority)]
-        internal void SendParametersUpdateRpc(ParametersUpdateMessage parametersUpdate)
+        [Rpc(SendTo.NotAuthority, AllowTargetOverride = true, InvokePermission = RpcInvokePermission.Owner)]
+        internal void SendParametersUpdateRpc(ParametersUpdateMessage parametersUpdate, RpcParams rpcParams = default)
         {
             m_NetworkAnimatorStateChangeHandler.ProcessParameterUpdate(parametersUpdate);
         }
@@ -1364,11 +1548,11 @@ namespace Unity.Netcode.Components
         /// <summary>
         /// Client-Server: Updates the client's animator's parameters
         /// </summary>
-        [ClientRpc]
-        internal void SendParametersUpdateClientRpc(ParametersUpdateMessage parametersUpdate, ClientRpcParams clientRpcParams = default)
+        [Rpc(SendTo.NotMe, AllowTargetOverride = true)]
+        internal void SendClientParametersUpdateRpc(ParametersUpdateMessage parametersUpdate, RpcParams rpcParams = default)
         {
             var isServerAuthoritative = IsServerAuthoritative();
-            if (!isServerAuthoritative && !IsOwner || isServerAuthoritative)
+            if ((!isServerAuthoritative && !IsOwner) || (isServerAuthoritative && !IsServer))
             {
                 m_NetworkAnimatorStateChangeHandler.ProcessParameterUpdate(parametersUpdate);
             }
@@ -1378,8 +1562,8 @@ namespace Unity.Netcode.Components
         /// Server-side animation state update request
         /// The server sets its local state and then forwards the message to the remaining clients
         /// </summary>
-        [ServerRpc]
-        private void SendAnimStateServerRpc(AnimationMessage animationMessage, ServerRpcParams serverRpcParams = default)
+        [Rpc(SendTo.Server, AllowTargetOverride = true)]
+        private void SendServerAnimStateRpc(AnimationMessage animationMessage, RpcParams rcParams = default)
         {
             if (IsServerAuthoritative())
             {
@@ -1387,7 +1571,7 @@ namespace Unity.Netcode.Components
             }
             else
             {
-                if (serverRpcParams.Receive.SenderClientId != OwnerClientId)
+                if (rcParams.Receive.SenderClientId != OwnerClientId)
                 {
                     return;
                 }
@@ -1403,25 +1587,26 @@ namespace Unity.Netcode.Components
                     return;
                 }
 
-                m_ClientSendList.Clear();
+                m_TargetGroup.Clear();
+
                 foreach (var clientId in connectedClientIds)
                 {
-                    if (clientId == serverRpcParams.Receive.SenderClientId || clientId == NetworkManager.ServerClientId || !NetworkObject.Observers.Contains(clientId))
+                    if (clientId == rcParams.Receive.SenderClientId || clientId == NetworkManager.ServerClientId || !NetworkObject.Observers.Contains(clientId))
                     {
                         continue;
                     }
-                    m_ClientSendList.Add(clientId);
+                    m_TargetGroup.Add(clientId);
                 }
-                m_ClientRpcParams.Send.TargetClientIds = m_ClientSendList;
-                m_NetworkAnimatorStateChangeHandler.SendAnimationUpdate(animationMessage, m_ClientRpcParams);
+                m_RpcParams.Send.Target = m_TargetGroup;
+                m_NetworkAnimatorStateChangeHandler.SendAnimationUpdate(animationMessage, m_RpcParams);
             }
         }
 
         /// <summary>
         /// Client-Server: Internally-called RPC client-side receiving function to update animation states
         /// </summary>
-        [ClientRpc]
-        internal void SendAnimStateClientRpc(AnimationMessage animationMessage, ClientRpcParams clientRpcParams = default)
+        [Rpc(SendTo.NotServer, AllowTargetOverride = true)]
+        internal void SendClientAnimStateRpc(AnimationMessage animationMessage, RpcParams rpcParams = default)
         {
             ProcessAnimStates(animationMessage);
         }
@@ -1429,12 +1614,16 @@ namespace Unity.Netcode.Components
         /// <summary>
         /// Distributed Authority: Internally-called RPC non-authority receiving function to update animation states
         /// </summary>
-        [Rpc(SendTo.NotAuthority)]
-        internal void SendAnimStateRpc(AnimationMessage animationMessage)
+        [Rpc(SendTo.NotAuthority, AllowTargetOverride = true, InvokePermission = RpcInvokePermission.Owner)]
+        internal void SendAnimStateRpc(AnimationMessage animationMessage, RpcParams rpcParams = default)
         {
             ProcessAnimStates(animationMessage);
         }
 
+        /// <summary>
+        /// Process incoming <see cref="AnimationMessage"/>.
+        /// </summary>
+        /// <param name="animationMessage">The message to process.</param>
         private void ProcessAnimStates(AnimationMessage animationMessage)
         {
             if (HasAuthority)
@@ -1454,17 +1643,15 @@ namespace Unity.Netcode.Components
             }
         }
 
-
-
         /// <summary>
         /// Server-side trigger state update request
         /// The server sets its local state and then forwards the message to the remaining clients
         /// </summary>
-        [ServerRpc]
-        internal void SendAnimTriggerServerRpc(AnimationTriggerMessage animationTriggerMessage, ServerRpcParams serverRpcParams = default)
+        [Rpc(SendTo.Server, AllowTargetOverride = true)]
+        internal void SendServerAnimTriggerRpc(AnimationTriggerMessage animationTriggerMessage, RpcParams rpcParams = default)
         {
             // Ignore if a non-owner sent this.
-            if (serverRpcParams.Receive.SenderClientId != OwnerClientId)
+            if (rpcParams.Receive.SenderClientId != OwnerClientId)
             {
                 if (m_LocalNetworkManager.LogLevel == LogLevel.Developer)
                 {
@@ -1478,23 +1665,22 @@ namespace Unity.Netcode.Components
 
             var connectedClientIds = m_LocalNetworkManager.ConnectionManager.ConnectedClientIds;
 
-            m_ClientSendList.Clear();
+            m_TargetGroup.Clear();
             foreach (var clientId in connectedClientIds)
             {
                 if (clientId == NetworkManager.ServerClientId || !NetworkObject.Observers.Contains(clientId))
                 {
                     continue;
                 }
-                m_ClientSendList.Add(clientId);
+                m_TargetGroup.Add(clientId);
             }
             if (IsServerAuthoritative())
             {
-                m_NetworkAnimatorStateChangeHandler.QueueTriggerUpdateToClient(animationTriggerMessage, m_ClientRpcParams);
+                m_NetworkAnimatorStateChangeHandler.QueueTriggerUpdateToClient(animationTriggerMessage, m_RpcParams);
             }
             else if (connectedClientIds.Count > (IsHost ? 2 : 1))
             {
-                m_ClientSendList.Remove(serverRpcParams.Receive.SenderClientId);
-                m_NetworkAnimatorStateChangeHandler.QueueTriggerUpdateToClient(animationTriggerMessage, m_ClientRpcParams);
+                m_NetworkAnimatorStateChangeHandler.QueueTriggerUpdateToClient(animationTriggerMessage, m_RpcParams);
             }
         }
 
@@ -1507,12 +1693,12 @@ namespace Unity.Netcode.Components
         }
 
         /// <summary>
-        /// Distributed Authority: Internally-called RPC client receiving function to update a trigger when the server wants to forward
-        ///  a trigger to a client
+        /// Distributed Authority: Internally-called RPC client receiving function to update a trigger when the authority wants
+        /// to forward a trigger to a client
         /// </summary>
         /// <param name="animationTriggerMessage">the payload containing the trigger data to apply</param>
-        [Rpc(SendTo.NotAuthority)]
-        internal void SendAnimTriggerRpc(AnimationTriggerMessage animationTriggerMessage)
+        [Rpc(SendTo.NotAuthority, AllowTargetOverride = true, InvokePermission = RpcInvokePermission.Owner)]
+        internal void SendAnimTriggerRpc(AnimationTriggerMessage animationTriggerMessage, RpcParams rpcParams = default)
         {
             InternalSetTrigger(animationTriggerMessage.Hash, animationTriggerMessage.IsTriggerSet);
         }
@@ -1523,8 +1709,8 @@ namespace Unity.Netcode.Components
         /// </summary>
         /// <param name="animationTriggerMessage">the payload containing the trigger data to apply</param>
         /// <param name="clientRpcParams">unused</param>
-        [ClientRpc]
-        internal void SendAnimTriggerClientRpc(AnimationTriggerMessage animationTriggerMessage, ClientRpcParams clientRpcParams = default)
+        [Rpc(SendTo.NotServer, AllowTargetOverride = true)]
+        internal void SendClientAnimTriggerRpc(AnimationTriggerMessage animationTriggerMessage, RpcParams rpcParams = default)
         {
             InternalSetTrigger(animationTriggerMessage.Hash, animationTriggerMessage.IsTriggerSet);
         }
@@ -1567,10 +1753,7 @@ namespace Unity.Netcode.Components
                 {
                     /// <see cref="UpdatePendingTriggerStates"/> as to why we queue
                     m_NetworkAnimatorStateChangeHandler.QueueTriggerUpdateToClient(animTriggerMessage);
-                    if (!IsHost)
-                    {
-                        InternalSetTrigger(hash, setTrigger);
-                    }
+                    InternalSetTrigger(hash, setTrigger);
                 }
                 else
                 {
@@ -1598,6 +1781,39 @@ namespace Unity.Netcode.Components
         public void ResetTrigger(int hash)
         {
             SetTrigger(hash, false);
+        }
+
+        /// <summary>
+        /// Allows for the enabling or disabling the synchronization of a specific <see cref="UnityEngine.Animator"/> parameter.
+        /// </summary>
+        /// <param name="parameterName">The <see cref="string"/> name of the parameter.</param>
+        /// <param name="isEnabled">Whether to enable or disable the synchronization of the parameter.</param>
+        public void EnableParameterSynchronization(string parameterName, bool isEnabled)
+        {
+            EnableParameterSynchronization(Animator.StringToHash(parameterName), isEnabled);
+        }
+
+        /// <summary>
+        /// Allows for the enabling or disabling the synchronization of a specific <see cref="UnityEngine.Animator"/> parameter.
+        /// </summary>
+        /// <param name="parameterNameHash">The hash value (from using <see cref="Animator.StringToHash(string)"/>) of the parameter name.</param>
+        /// <param name="isEnabled">Whether to enable or disable the synchronization of the parameter.</param>
+        public void EnableParameterSynchronization(int parameterNameHash, bool isEnabled)
+        {
+            var serverAuthoritative = OnIsServerAuthoritative();
+            if (!IsSpawned || serverAuthoritative && IsServer || !serverAuthoritative && IsOwner)
+            {
+                for (int i = 0; i < m_CachedAnimatorParameters.Length; i++)
+                {
+                    var cachedParameter = m_CachedAnimatorParameters[i];
+                    if (cachedParameter.Hash == parameterNameHash)
+                    {
+                        cachedParameter.Exclude = !isEnabled;
+                        m_CachedAnimatorParameters[i] = cachedParameter;
+                        break;
+                    }
+                }
+            }
         }
     }
 }